+++ src/ImgixImageStyleFactoryInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Drupal\imgix;
+
+use Drupal\image\ImageStyleInterface;
+
+interface ImgixImageStyleFactoryInterface
+{
+    public function getImageStyleByParameters(array $parameters): ImageStyleInterface;
+}